Fix question about NSArray and NSSet
I don't actually know if there are different performance characteristics in terms of _iterating through_ NSSet or NSArray, but reading the given answer, that's not really what this question is about. It's asking how quickly a given item can be found. I've updated the question and answer to make this clearer.

For the original question, I'd actually guess that NSArray would be faster as it should be more densely packed in memory, whereas sets are typically sparse, so at the very least it would be difficult to take advantage of CPU caches when iterating through.